requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
joblib==1.3.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.0
wrapt==1.15.0
zipp==3.17.0
